install yarn

+++ start.sh
@@ -13,8 +13,12 @@ echo "Cloning into $FN"
 git clone $REPO 
 cd $FN
 git checkout $REV
+if [[ -f yarn.lock ]];then
+    npm install -g yarn
+else
+    npm install npm@8
+fi
 
-npm install npm@8
 export PATH=$(pwd)/node_modules/.bin:$PATH
 #npm install --legacy-peer-deps
